+++ src/3_game_loop/a_game_logic/hooks/loop_hook.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   loop_hook.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fschuber <fschuber@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jkauker <jkauker@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 10:23:28 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/14 12:07:49 by fschuber         ###   ########.fr       */
+/*   Updated: 2024/06/19 16:03:23 by jkauker          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,6 +35,7 @@ void	loop_hook(void *param)
 	{
 		render_game_scene();
 		handle_mouse_mv();
+		handle_player_move();
 		game()->prev_time = curr_time;
 	}
 }